split up services and UI
